Added more classes for concepts.
There is now a skeleton for a concept factory and an abstract class which
implements all mechanisms to break the intents and extents into the different
contingents.

The factory is started as a class instead of an interface since the schema
has to be given in both cases (memory-mapped and DB connection) and
the Concepts will handle the abstraction of the storage version.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@38 f8b840a1-a80d-0410-987b-95fbfcb7d1ee