Refactoring drawing code.
We now have some concept of canvas items drawn in the diagram which
draw themself using a special drawing context which scales only points but
not shapes. These items are stored in DiagramView as list and accessed
mostly through the interface.

This is close to a generic canvas implementation which should be the next
refactoring step.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@42 f8b840a1-a80d-0410-987b-95fbfcb7d1ee