Further refactoring for introducing DiagramNodes.
The DiagramNode class now has the LabelInfos attached. These contain still
the contingents, the Concept interface is not yet used. Otherwise the
refactoring for the explicit DiagramNodes should be nearly finished.

The getPoint() and similar functions where renamed to getPosition() and
similar names to avoid misunderstanding -- position is clearly distinguishable
from node, point is not.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@51 f8b840a1-a80d-0410-987b-95fbfcb7d1ee