The Concept interface and its implementations are now used.
DiagramNode now stores a Concept, LabelInfo is abstract, its subclasses
AttributeLabelInfo and ObjectLabelInfo use the appropriate contingent
iterators to access the entries. The access on the entries of the labels has
been changed, too. The parser now created nodes explicitely and lines
from a node to a node (instead of using indices), but it still silently assumes
to have to create memory-mapped Concepts, the DB version still has to be
done.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@53 f8b840a1-a80d-0410-987b-95fbfcb7d1ee