Database connection is now happening.
There is now an implementation for connecting Concepts to an ODBC
database. It seems to work quite well for non-nested, non-filtered diagrams,
although the PCTest example we got was quite broken -- some diagrams
still fail in there and we probably want to print out error messages whenever
we notice failures. Maybe a test mode to check for duplicates (same object
in multiple contingents) would be useful, too.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@54 f8b840a1-a80d-0410-987b-95fbfcb7d1ee