Checkin of a bunch of new code written over the weekend.
This checkin includes a bunch of new code written over the weekend. The
interface for sets is now implemented and work as begun on the interface
for relations.

This first implementation uses the STL and is intened as a proof of concept for the interface. Later implementations will be more efficient.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@62 f8b840a1-a80d-0410-987b-95fbfcb7d1ee