Conversion to use a single C++ interface.
A significant effort has been put in to allow generation of python and
java interfaces from a single C++ source.

There are still a number of unsatisfactory elements such as the lack
of namespaces, the lack of support for native clone mechanisms such as
clone() in java and copy in python.

The interface is however still evolving so perhaps these things can be
tollerated in the short term.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@65 f8b840a1-a80d-0410-987b-95fbfcb7d1ee