Some fixes to Makefile.
Removed setting of variable in Makefile to make config.mak work. Changed some dependencies to create lib and lib_d when needed. Added current directory to LD_LIBRARY_PATH for the sarlcmodule. Renamed sarl.so to libsarl.so for java example.

Result: Sarl is compiling and testing fine on plueschix's machine, I still have trouble with the python example, but this may be due to my strange swig installation (swig is dumb when you don't install it into its default directories :-(, configure-switches aren't working for me).


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@83 f8b840a1-a80d-0410-987b-95fbfcb7d1ee