Instance types are now handled, links can be moved with attached vertices.
The instance types are checked when restricting types of nodes
and merging nodes, avoiding operations which cause the node to
have a type not compatible with the instance type.

When moving links the shift key can be used to move it
with all nodes attached, the ctrl key moves the whole graph
component (i.e. all vertices connected to it). Shortcuts and
esp. the default behaviour might be changed later on.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@95 f8b840a1-a80d-0410-987b-95fbfcb7d1ee