Added treeviews for the canon.
There are now three additional panels on the right hand
side of Crepe, showing the type hierarchy, the relation
hierarchies and the list of instances.

To update these views the KnowledgeBase emits events
notifying the addition of types, relations and instances.
The type and instances views do display data, although
they do not yet have operations. The relations are not
yet displayed.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@113 f8b840a1-a80d-0410-987b-95fbfcb7d1ee