Added support for exporting PDF from the canvas.
Thanks to another open source library and the support of one of their
project leaders, we are now able to export PDF graphic from the canvas.

Unfortunately we have to do this as pure graphics, since exporting text
as text did result in problems with font metrics (as usual). But this way
the results are very accurate in respect to the look on screen.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@237 f8b840a1-a80d-0410-987b-95fbfcb7d1ee