Changed the way the list is filled.
Changing the model step by step did cause performance problems, I
suspect that the view got notified on each insertion. Sometimes this
whole process seemed even to break.

Now we create an array for the extent and display that. Somehow our
Concept interface is not really suited for what we are doing here. It sure
needs refactoring.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@357 f8b840a1-a80d-0410-987b-95fbfcb7d1ee