Added support for negation and enhanced labels.
(a) we support negation now. As long as at least one clause is positive,
negation can be achieved by combining all positive clauses with the 
negative one
(b) the labels don't contain the "body:" parts anymore
(c) we forget the required parts (the "+"). For the normal queries it
doesn't make any difference since we query only single clauses. For the
negations it does cause new implications which we don't want


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@381 f8b840a1-a80d-0410-987b-95fbfcb7d1ee