Added support for mouse enter/exit events.
The former CanvasItemPointedEvent has been split into two events for
mouse enter/exit. These events are sent whenever the mouse is over a
new canvas item -- one event for the item left, one for the item entered.

There is a new method on the canvas controller to hide the mouse from 
the items, which causes these events to be sent, too. On a hide call the
mouse exit event will be sent about the current item, on a show call the
opposite event is sent.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@403 f8b840a1-a80d-0410-987b-95fbfcb7d1ee