Still running into the broken index problem from time to time. Next fix.
My suspicion is that this is caused by some things running on the index
while we are deleting it. Now the code gets rid of the old query event
handler before deletion of the index. But I think we have to check if
there is something we are meant to do to shut the query engine down
properly.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@423 f8b840a1-a80d-0410-987b-95fbfcb7d1ee