Moved all index-related code into a single class.
This includes a change to a 1:1 relation between indexes and indexed
directories (made so we can store relative locations later). At the
moment this reduces the capabilities since it means that we can index
only one directory. Should be easy to fix, though.

On the positive side the updates are now handled in a background 
thread and the indexing code is simplified -- no synchronization needed
and no distinction between creating an index or updating one (creating
is an update on an empty index).


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@551 f8b840a1-a80d-0410-987b-95fbfcb7d1ee