Fixed class loader.
The problem was that the jar URLs had a leading slash for the jar files.
Maybe there were other issues, I removed some regexp replacements
that seemed inappropriate. I also added some more logging code. In
the end I think it was only the little / that caused the problem, but after
all that's what debugging is about :-)


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@774 f8b840a1-a80d-0410-987b-95fbfcb7d1ee