Changed bahaviour of ListSet.add(...) methods.
The old way of throwing exceptions when adding an object which has been
in the collection before got a bit annoying. The new approach implements
Set.add(Object) and a variation of List.add(int,Object), where the specs
of List.add(..) are not obeyed in the case the object is already in the
collection.

The differences are documented on the Interface, though.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@809 f8b840a1-a80d-0410-987b-95fbfcb7d1ee