Docco now keeps a history of the last 10 queries made, offered as combo box.
After a quite painful process of trial and error I figured out how to use editable
combo boxes in Swing. This is now used to replace the query field with such a
thing, where you can either enter a new query or pick one of the last ones
(up to ten).

The history of the last queries is stored in the session management and restored
on startup. Since I couldn't figure how to get an empty text field with that approach
I went the other way and let Docco requery the last query from the last session
instead.


git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@821 f8b840a1-a80d-0410-987b-95fbfcb7d1ee