Added some basic support for Analyzers.
Analyzers in Lucene are code parts that model tokenization, filtering and stemming. The indexing process now supports a selection of those, which are available from the Lucene distribution.

This change has the effect that old indexes won't work anymore since the .contents file has been replaced with a .properties file. Reindexing is required.

The queries are not yet adjusted, they always use the StandardAnalyzer. That means query terms are not stemmed, which causes oddities if an index uses anything but the StandardAnalyzer. To be fixed before release, but as long as the analyzer is not changed it should behave as before.

Another problem I encountered is that the updates don't work anymore. I don't think that is related to this commit, though -- it is probably one of the changes in date handling coming from the Lucene upgrade.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@966 f8b840a1-a80d-0410-987b-95fbfcb7d1ee