Finalized support for analyzers.
The queries now use the analyzers, too -- this means all query terms get stemmed. This happens on a per-index basis.

To use different analyzers for each index, the code for searching had to be changed a bit. During this the extra set interface and class got replaced with an array -- less code and potentially a bit faster, althought that would be in the neglectable range since the N in the complexity formula would be the number of term.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@967 f8b840a1-a80d-0410-987b-95fbfcb7d1ee