Got rid of all Eclipse warnings.
All warnings Eclipse generates in my current setup have been removed. Most of them where in the Multivalent source, where I removed them through quite violent means, but since I was quite confident the removed code wasn't used and the code base is not meant to be maintained, I didn't really hesitate :-)

Maybe we should see if newer Multivalent versions can be used as a normal library, the old one had the problem that it did System.exit()s whenever it was unhappy with a file.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@982 f8b840a1-a80d-0410-987b-95fbfcb7d1ee