Changed help and message loading mechanism.
Both messages and help contents are now loaded from subdirectories of the "resources" folder. That sorts things a bit nicer and it allows loading the help in a way that works from within the IDE, too.

Help is also grouped in a language-specific subfolder and the language-neutral images (logos). More subfolders can be created to have localized help, the access path from the menu is in the resource bundles along with the messages.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1001 f8b840a1-a80d-0410-987b-95fbfcb7d1ee