Added basic inference capabilities.
This uses Jena's generic rule engine. The rules are expressed in a separate file with reasonable syntax, which should make things easier to maintain. RDFS/OWL would be an option, too -- I haven't really explored that one yet. Since we basically just use unions and transitive closures it should be easy to express our rules that way.

One problem is that Jena doesn't do inference in the persisted model, which means the whole inferred model has to fit into memory to be added to the persisted one later on. This could cause trouble with scaling to larger source bases.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1049 f8b840a1-a80d-0410-987b-95fbfcb7d1ee