Export now runs as Eclipse Job.
Which means it can run in the background, can be cancelled and gives a little bit of indication about what it is doing (currently only two states: parsing code, inferring relations).

It automatically gets scheduled as user job (i.e. pops up as dialog when using standard settings) and with a lock on the workspace. Locking the workspace is a little overkill in some regards but as long as we do not distinguish the project in the export settings it is the only one that is safe.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1056 f8b840a1-a80d-0410-987b-95fbfcb7d1ee