Changed output format to RDF file instead of database.
The output is now written into a standard RDF XML file. This is more flexible and since the model is kept in memory without extra layers involved, more faster, too. The ToscanaJ export is down from ~35min to ~7min, and it runs on a standard Eclipse configuration without adjusting memory limits.

To write RDF the URIs had to be fixed to follow the official syntax, i.e. we had to escape some characters.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1060 f8b840a1-a80d-0410-987b-95fbfcb7d1ee