Added support for ignore patterns.
During the recursion elements can be skipped by providing regular expressions that are matched against the element names. If one of the expressions provided matches, the element and all children will be ignored.

Currently the list is still hard-coded due to a lack of a GUI-component for it. The default excludes JARs, and everything ending on ".test" or ".tests".

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1072 f8b840a1-a80d-0410-987b-95fbfcb7d1ee