Changed type parameters for the event broker.
The old version was too strict, certain things that should have worked didn't. Unfortunately a proper solution to this isn't that easy, since the whole typing is getting a bit messy. I think the best solution would be deprecating the EventBroker in favour of a simpler approach such as an observer pattern. AFAIK it still isn't threadsafe either.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1110 f8b840a1-a80d-0410-987b-95fbfcb7d1ee