Removed some commented code.
This code would have added extra metadata PDF provides but Docco doesn't support yet. If anyone will ever want that we'll find the ways to get to the data quickly (it is easy enough) and for now it just makes things a bit more messy, so I removed it all.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1114 f8b840a1-a80d-0410-987b-95fbfcb7d1ee