Removed PDF from the default mappings for Multivalent.
Multivalent's PDF support isn't particularly good (lots of documents fail) and it collides in installations where PDFbox is available as alternatives -- if Multivalent is registered first, it will get the documents instead of PDFbox.

By removing the PDF extension from the default filter Multivalent will not try to process PDF documents by default, but it can still be used if this mapping is created explicitly by the user.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1120 f8b840a1-a80d-0410-987b-95fbfcb7d1ee