Updated POI library to 3.2final.
Just the library change. I decided to use the library version number in this change (and further changes), since it is more obvious and it doesn't really matter much if there's a coherent history in the version control system -- with Subversion it's easy enough to look at the history of the directory instead.

git-svn-id: svn://svn.code.sf.net/p/tockit/code/trunk@1121 f8b840a1-a80d-0410-987b-95fbfcb7d1ee